Extract embed-image in separate file
